+++ subsys/bluetooth/host/hci_core.c
@@ -658,60 +658,6 @@ static int set_adv_random_address(struct bt_le_ext_adv *adv,
 	adv->random_addr.type = BT_ADDR_LE_RANDOM;
 	return 0;
 }
-
-int bt_addr_from_str(const char *str, bt_addr_t *addr)
-{
-	int i, j;
-	uint8_t tmp;
-
-	if (strlen(str) != 17U) {
-		return -EINVAL;
-	}
-
-	for (i = 5, j = 1; *str != '\0'; str++, j++) {
-		if (!(j % 3) && (*str != ':')) {
-			return -EINVAL;
-		} else if (*str == ':') {
-			i--;
-			continue;
-		}
-
-		addr->val[i] = addr->val[i] << 4;
-
-		if (char2hex(*str, &tmp) < 0) {
-			return -EINVAL;
-		}
-
-		addr->val[i] |= tmp;
-	}
-
-	return 0;
-}
-
-int bt_addr_le_from_str(const char *str, const char *type, bt_addr_le_t *addr)
-{
-	int err;
-
-	err = bt_addr_from_str(str, &addr->a);
-	if (err < 0) {
-		return err;
-	}
-
-	if (!strcmp(type, "public") || !strcmp(type, "(public)")) {
-		addr->type = BT_ADDR_LE_PUBLIC;
-	} else if (!strcmp(type, "random") || !strcmp(type, "(random)")) {
-		addr->type = BT_ADDR_LE_RANDOM;
-	} else if (!strcmp(type, "public-id") || !strcmp(type, "(public-id)")) {
-		addr->type = BT_ADDR_LE_PUBLIC_ID;
-	} else if (!strcmp(type, "random-id") || !strcmp(type, "(random-id)")) {
-		addr->type = BT_ADDR_LE_RANDOM_ID;
-	} else {
-		return -EINVAL;
-	}
-
-	return 0;
-}
-
 static void adv_rpa_invalidate(struct bt_le_ext_adv *adv, void *data)
 {
 	if (!atomic_test_bit(adv->flags, BT_ADV_LIMITED)) {
@@ -824,7 +770,7 @@ static int le_set_private_addr(uint8_t id)
 		return err;
 	}
 
-	nrpa.val[5] &= 0x3f;
+	BT_ADDR_SET_NRPA(&nrpa);
 
 	return set_random_address(&nrpa);
 }
@@ -839,7 +785,7 @@ static int le_adv_set_private_addr(struct bt_le_ext_adv *adv)
 		return err;
 	}
 
-	nrpa.val[5] &= 0x3f;
+	BT_ADDR_SET_NRPA(&nrpa);
 
 	return set_adv_random_address(adv, &nrpa);
 }
@@ -6082,41 +6028,6 @@ static int set_event_mask(void)
 
 	sys_put_le64(mask, ev->events);
 	return bt_hci_cmd_send_sync(BT_HCI_OP_SET_EVENT_MASK, buf, NULL);
-}
-
-static inline int create_random_addr(bt_addr_le_t *addr)
-{
-	addr->type = BT_ADDR_LE_RANDOM;
-
-	return bt_rand(addr->a.val, 6);
-}
-
-int bt_addr_le_create_nrpa(bt_addr_le_t *addr)
-{
-	int err;
-
-	err = create_random_addr(addr);
-	if (err) {
-		return err;
-	}
-
-	BT_ADDR_SET_NRPA(&addr->a);
-
-	return 0;
-}
-
-int bt_addr_le_create_static(bt_addr_le_t *addr)
-{
-	int err;
-
-	err = create_random_addr(addr);
-	if (err) {
-		return err;
-	}
-
-	BT_ADDR_SET_STATIC(&addr->a);
-
-	return 0;
 }
 
 static uint8_t bt_read_public_addr(bt_addr_le_t *addr)
